refactor ngen request class into module
